fix(lua): lcd.drawGauge wide gauge error (#2375)

+++ radio/src/lua/api_colorlcd.cpp
@@ -954,7 +954,7 @@ static int luaLcdDrawGauge(lua_State *L)
   flags = flagsRGB(flags);
   
   luaLcdBuffer->drawRect(x, y, w, h, 1, 0xff, flags);
-  uint8_t len = limit((uint8_t)1, uint8_t(w*num/den), uint8_t(w));
+  uint16_t len = limit((uint16_t)1, uint16_t(w*num/den), uint16_t(w));
   luaLcdBuffer->drawSolidFilledRect(x+1, y+1, len, h-2, flags);
 
   return 0;
